requirements.txt
requests
matplotlib
pandas
mysql-connector-python
tenacity
requests
numpy
matplotlib
